requirements.txt
hyperopt==0.2.7
mlflow==1.26.1
numpy==1.21.5
pandas==1.3.4
scikit_learn==1.2.1
scikit_plot==0.3.7
xgboost==1.7.1
Flask==2.2.3
Flask-Cors==4.0.0
Werkzeug==2.2.2
#AUTH0 requirements
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
#MONGO
pymongo==4.8.0
#BENCHMARK
psutil